refactor: move and rename Dashboard page to Courses page
