+++ apps/web/routes.ts
@@ -14,7 +14,7 @@ export const routes: (
       route("", "modules/Statistics/Statistics.page.tsx", {
         index: true,
       });
-      route("courses", "modules/Dashboard/Dashboard.page.tsx");
+      route("courses", "modules/Courses/Courses.page.tsx");
       route("course/:id", "modules/Courses/NewCourseView/NewCourseView.page.tsx");
       route("course/:courseId/lesson/:lessonId", "modules/Courses/Lesson/Lesson.page.tsx");
       route("settings", "modules/Dashboard/Settings/Settings.page.tsx");
